Organise code that operates on units and their utility functions.
