+++ src/main/webapp/ui/src/stores/definitions/__tests__/Units/truncateIsoTimestamp.test.js
@@ -4,7 +4,7 @@
 //@flow
 /* eslint-env jest */
 import "@testing-library/jest-dom";
-import { truncateIsoTimestamp } from "../../conversions";
+import { truncateIsoTimestamp } from "../../Units";
 
 describe("truncateIsoTimestamp", () => {
   test("Simple examples with string", () => {
